+++ client/src/pages/Auth/Register.js
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 import { Link } from "react-router-dom";
 import { signup } from "../../_actions/userAction";
 
-import classes from "./Register.module.css";
+import classes from "../../styles/Register.module.css";
 import { FaArrowLeft } from "react-icons/fa";
 
 const Register = (props) => {
